fix: move ConsumptionOverlayError into exceptions.py, subclass BESSException
Review finding (round 1, PR #705): every other custom exception in this
codebase lives in core/bess/exceptions.py and inherits BESSException;
ConsumptionOverlayError was the only exception type defined outside
that file, subclassing bare Exception. Per rules.md's Error Handling
section and .github/claude-bot.md's checklist. Moved the class,
updated every import site to pull it from exceptions.py directly.

Refs #428
